Fix issue when only finished tasks were showing
Unprepared query id that is used to retrieve unfinished tasks
was ovewritten by sequential data request thus resulting in
failure to check for valid query id when handling records.
As a result, handler remained in the queue and was popped
to call callback from wrong caller.
Multiple sequential requests resulted in piling up handlers
in the queue.
To fix this issue, query ids were also added to their own queue
and test was written to demonstrate faulty behavior.
In addition, it turned out that there was extra request for data
that was not needed (and it was causing behavior described above).
This extra request was removed by fixing signal connection.